create screen width updater hook
