+++ src/components/NavBar/NavBar.jsx
@@ -6,6 +6,7 @@ import {
 } from 'react';
 import { Link } from 'react-router-dom';
 import { Sling as Hamburger } from 'hamburger-react';
+import { useScreenWidth } from '../customHooks/useScreenWidth';
 import navStyle from './NavBar.module.css';
 
 // files
@@ -13,17 +14,10 @@ import ResumeFile from '../../pdf/sethvm_resume.pdf';
 
 export default function NavBar() {
 
-    const [ screenWidth, setScreenWidth ] = useState(window.screen.width);
+    const screenWidth = useScreenWidth();
     const [ expanded, setExpanded ] = useState(false);
 
-    const onDesktop = (screenWidth >= 992);
-
     const toggleButton = createRef();
-
-    const updateScreenWidth = () => {
-        const newWidth = window.screen.width;
-        setScreenWidth(newWidth);
-    }
 
     const toggleNav = () => {
         setTimeout(() => { setExpanded(!expanded) }, 160);
@@ -32,14 +26,6 @@ export default function NavBar() {
     const closeNav = useCallback(() => {
         setTimeout(() => { setExpanded(false) }, 160);
     }, []);
-
-    // screen width change listener
-    useEffect(() => {
-        window.addEventListener('resize', updateScreenWidth);
-        return () => {
-            window.removeEventListener('resize', updateScreenWidth);
-        }
-    }, [])
 
     // close overlay on outside click if open
     const handleClick = useCallback((e) => {
@@ -108,7 +94,7 @@ export default function NavBar() {
                     </li>
                 </ul>
             </div>
-            {onDesktop &&
+            {(screenWidth >= 992) &&
                 <Link
                 aria-label='About'
                 to='/about'>
